+++ libs/str.h
@@ -5,3 +5,5 @@ char *strcaseswap(char *str, ssize_t len);
 char *vstrjoint(char *dst, const char *fmt, ...);
 const char *vstrcat(int nstr, ...);
 const char *vstrcat_r(char buf[], int nstr, ...);
+
+unsigned long str2size(const char *str);