+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 - glibc: test `__GLIBC_USE()`;
 - shell: add some commands from ostools;
 #### Changed
+- Move str2size() from `oom_helpers.c` to libs/str.c;
 - libs: Rename macro `TEST` to `TEST_MAIN`;
 - pthread: use fhexdump() instead of memshow();
 - Move hexdump() from fakeCUDA to libs/memshow.c;